backend/requirements.txt
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
emoji==2.11.0
exceptiongroup==1.2.0
fake-useragent==1.5.1
filetype==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
python-dotenv==1.0.1
regex==2023.12.25
sniffio==1.3.1
tqdm==4.66.2
twikit==1.4.4
typing_extensions==4.11.0
